updates to randoop coverage tools/scripts

+++ framework/test/randoop_single_coverage.sh
@@ -1,12 +1,15 @@
 #!/usr/bin/env bash
 ################################################################################
 #
-# This script generates coverage data for Randoop generated tests over the defects4j suite.
+# This script generates coverage data for Randoop generated tests over a single defects4j test.
 #
 ################################################################################
 # Import helper subroutines and variables, and init Defects4J
 source test.include
 init
+
+# change 0 to 1 for extra logging info and to save temp files
+debug=0
 
 # master coverage file
 master_coverage=$TMP_DIR/coverage
@@ -14,10 +17,10 @@ master_coverage=$TMP_DIR/coverage
 # Directory for Randoop test suites
 randoop_dir=$TMP_DIR/randoop
 
-# Generate tests for all projects
+# Generate tests for Lang-4
 # Mockito #1 and #3 don't work; problem finding classes in byte-buddy?
 # Thus the strange hack below
-projects=( Chart Closure Lang Math Mockito Time )
+projects=( Lang )
 type=f
 
 # Test suite source and number
@@ -28,14 +31,18 @@ suite_num=1
 #rm -f $master_coverage
 
 for pid in "${projects[@]}"; do
-    for bid in 1 2 3 4 5; do
+    for bid in 4; do
         if [ "$pid" = 'Mockito' ]; then
             bid=$(($bid + 3))
         fi
         vid=${bid}$type
 
         # Run Randoop
-        run_randoop.pl -p $pid -v $vid -n 1 -o $randoop_dir -b 100 || die "run Randoop on $pid-$vid"
+        if [ "$debug" = '0' ]; then
+            run_randoop.pl    -p $pid -v $vid -n 1 -o $randoop_dir -b 100 || die "run Randoop on $pid-$vid"
+        else
+            run_randoop.pl -D -p $pid -v $vid -n 1 -o $randoop_dir -b 100 || die "run Randoop on $pid-$vid"
+        fi
     done
 
     suite_dir=$randoop_dir/$pid/$suite_src/$suite_num
@@ -48,4 +55,6 @@ for pid in "${projects[@]}"; do
 done
 
 # delete tmp file directory
-rm -rf $randoop_dir
+if [ "$debug" = '0' ]; then
+    rm -rf $randoop_dir
+fi
